requirements.txt
requests==2.28.1
pandas==1.5.0
python-dotenv==0.19.2
cryptography==36.0.1
